Refactor insertParameters to its own module, and convert to TypeScript
